fix: add types folder and move there
